Script to get the target build
The target build means:
The latest release build today. If not, it will be the latest dev
build.
With the target build function, CI can do more meaningful task.
It means release builds and dev builds are both ready today, CI would
ignore the dev build and choose the release builds to do the environment
preparation and testing.
It is consistent with the function of the post-merge CI.
If the release branch is updated, the post-merge CI would ignore the
change of develop, work with the release branch and do testing.